pyproject.toml
[project]
name = "weather-bike-demo"
version = "0.1.0"
description = "Weather-aware bike demand platform demonstrating modern data stack with incremental data source value"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "altair>=4.2.2",
    "apache-airflow>=3.1.1",
    "apache-airflow-providers-common-sql>=1.28.2",
    "dbt-duckdb>=1.9.6",
    "dlt[duckdb]>=1.18.1",
    "great-expectations>=1.8.1",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.7",
    "pandas>=2.3.3",
    "plotly>=6.3.1",
    "polars>=1.35.1",
    "requests>=2.32.5",
    "seaborn>=0.13.2",
    "streamlit>=1.51.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "A", "C4", "T20", "SIM", "ARG"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["dlt_pipeline", "streamlit_app"]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-mock>=3.15.1",
    "ruff>=0.14.3",
]
